requirements.txt
pymupdf
tika
tqdm
python-docx
openpyxl
